refactor: remove unnesessary part of names for files
