+++ src/components/pages/EmployeesPage/EmployeesAdditionalButtons/EmployeesAdditionalButtons.tsx
@@ -8,10 +8,10 @@ import { RoutePath } from '../../../../constants/routeVariables';
 import { UserRoles } from '../../../../constants/userRoles';
 import { useUser } from '../../../../hooks/useUser';
 import { IAdditionalButtonsProps } from '../../../Table/TableRows/TableRowComponent.types';
-import { updateCacheAfterDeleteUser } from '../../../../graphql/cache/deleteUser.cache';
+import { updateCacheAfterDeleteUser } from '../../../../graphql/cache/deleteUser';
 import { DELETE_USER } from '../../../../graphql/mutations/deleteUser';
-import { DeleteUserResult } from '../../../../graphql/types/results/userTypeResult';
-import { DeleteUserInput } from '../../../../graphql/types/inputs/userTypeInput';
+import { DeleteUserResult } from '../../../../graphql/types/results/user';
+import { DeleteUserInput } from '../../../../graphql/types/inputs/user';
 import * as Styled from './EmployeesAdditionalButtons.styles';
 
 export const EmployeesAdditionalButtons: FC<IAdditionalButtonsProps> = ({ item }) => {
