+++ src/components/pages/EmployeesProfilePage/AvatarModal/AvatarModal.tsx
@@ -4,15 +4,15 @@ import { useForm } from 'react-hook-form';
 import { useMutation, useQuery } from '@apollo/client';
 import CancelIcon from '@mui/icons-material/Cancel';
 import { USER } from '../../../../graphql/queries/user';
-import { IAvatarReturn } from '../../../../graphql/types/results/avatarTypeResult';
+import { IAvatarReturn } from '../../../../graphql/types/results/avatar';
 import { DELETE_AVATAR, UPLOAD_AVATAR } from '../../../../graphql/mutations/avatar';
-import { updateUserCacheAfterAvatarMutation } from '../../../../graphql/cache/avatar.cache';
+import { updateUserCacheAfterAvatarMutation } from '../../../../graphql/cache/avatar';
 import { avatarSchema } from '../../../../utils/validationSchema';
 import { TError } from '../../../../types/errorTypes';
 import { convertToBase64 } from '../helpers/convertToBase64';
 import { ModalWindow } from '../../../UI/ModalWindow';
 import { Spinner } from '../../../Spinner';
-import { IUserAllResult } from '../../../../graphql/types/results/userTypeResult';
+import { IUserAllResult } from '../../../../graphql/types/results/user';
 import * as Styled from './AvatarModal.styles';
 import { InputFile } from './InputFile/InputFile';
 import { IAvatarForm, IAvatarModal } from './AvatarModal.types';
