+++ src/graphql/types/results/avatar.ts
