+++ src/graphql/types/results/cv.ts
