+++ src/graphql/types/results/department.ts
